backend/requirements.txt
Django==3.2.16
requests==2.26.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
djoser==2.1.0
PyJWT==2.1.0
django-filter==22.1
pytz==2022.1
python-dotenv==0.20.0
drf-extra-fields==3.4.1
drf-base64==2.0
isort==5.11.4
Pillow==9.5.0
webcolors==1.13
reportlab==4.0.0
